Get session duration using a function
The usage of Viper defaults isn't suitable in this case. Instead of
chaining a series of defaults (and then immediately using them) it is
much more intuitive and simple to encapsulate this logic in a
function. It also makes this logic testable.
The only thing I would consider using a Viper default for in this
case is the hardcoded value (3600), however even this seems
unnecessary since this value is hardcoded on AWS and is unlikely to
change often so there is no real reason to even make this
user-configurable.