Okta Challenge Preferences
As I had issues with Clisso not selecting push notifications but require me to enter codes instead of using push, even though I only had okta push turned on, I added this functionality, so we prefer push when possible but make it configurable for whoever is into doing it. Worked well here so far in different scenarios.